inode2filename: Fix a bug on parsing lines for inode2filename.
Some temp file has '\n' in its name, which breaks the parsing.
To solve this issue, add a rest line length in the begining as a 4 bytes
array. For example, suppose the line is "0x000x000x10K 253:9:7 ./test",
"0x000x000x0000x10" (16) is the length of "K 253:9:7 ./test".
Each line has no '\n' in the end and the end is told by the length.

Bug: 158488826
Test: ioraiorap.inode2filename -pm out --all -of textcache -v
Test: atest iorapd-tests
Change-Id: Id950e7e0ba3e5f76b336ba95da79d0279b0bf472